Transfers all contents of intake.csv module into db_utils as that module consisted of functions related to database upload and were closely linked and reliant on db functions / classes.
